back/requirements.txt
Django==3.1.4
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
django-filter==2.3.0